Setup the controller for GET API
